+++ produtos/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, get_object_or_404, redirect
+from django.contrib.auth.decorators import login_required
 from .models import Produto, Categoria
 from django import forms
 
@@ -9,15 +10,18 @@ class ProdutoForm(forms.ModelForm):
         fields = ['nome', 'descricao', 'preco', 'tipo_preco', 'categoria', 'disponivel']
 
 
+@login_required
 def home(request):
     return render(request, 'produtos/home.html')
 
 
+@login_required
 def lista_produtos(request):
     produtos = Produto.objects.all()
     return render(request, 'produtos/lista_produtos.html', {'produtos': produtos})
 
 
+@login_required
 def adicionar_produto(request):
     categorias = Categoria.objects.all()
     form = ProdutoForm()
@@ -32,6 +36,7 @@ def adicionar_produto(request):
     })
 
 
+@login_required
 def editar_produto(request, id):
     produto = get_object_or_404(Produto, id=id)
     categorias = Categoria.objects.all()
@@ -47,7 +52,10 @@ def editar_produto(request, id):
     })
 
 
+@login_required
 def excluir_produto(request, id):
+    if not request.user.is_staff:
+        return redirect('/produtos/')
     produto = get_object_or_404(Produto, id=id)
     if request.method == 'POST':
         produto.delete()
